Asignacion

+++ Caba/src/main/java/com/pagina/Caba/repository/AsignacionRepository.java
@@ -3,6 +3,7 @@ package com.pagina.Caba.repository;
 import com.pagina.Caba.model.Asignacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AsignacionRepository extends JpaRepository<Asignacion, Integer> {
+public interface AsignacionRepository extends JpaRepository<Asignacion, Long> {
 }
+
